Move the border glyph tables into ansi, and symbols to their classes
BorderEdgeStyle, BOX_DRAWING, and ROUNDED_CORNERS had one consumer,
the ansi module; they move in and stop being exported. Every
module-local symbol declaration moves from the top of its file to
just above the first statement that uses it, so a class and its
symbol keys read together.
