+++ src/internal/layout.ts
@@ -43,100 +43,6 @@ import {
 	writeClusterWidths,
 } from "./text.js";
 
-const kBoxes = Symbol("boxes");
-const kAnonymousBoxes = Symbol("anonymousBoxes");
-const kMeasureNodes = Symbol("measureNodes");
-const kTerminalReordersText = Symbol("terminalReordersText");
-const kDOMNodeByFlexNode = Symbol("domNodeByFlexNode");
-const kInvalidatedNodes = Symbol("invalidatedNodes");
-const kAddNode = Symbol("addNode");
-const kLayoutPass = Symbol("layoutPass");
-const kApplyRestyles = Symbol("applyRestyles");
-const kDirtyRunContainers = Symbol("dirtyRunContainers");
-const kPruneDisconnectedNodes = Symbol("pruneDisconnectedNodes");
-const kContainerFlexNode = Symbol("containerFlexNode");
-const kSyncContainerRuns = Symbol("syncContainerRuns");
-const kIsNodeLive = Symbol("isNodeLive");
-const kRetireAnonymousBox = Symbol("retireAnonymousBox");
-const kUntrackNode = Symbol("untrackNode");
-const kRunFlexNode = Symbol("runFlexNode");
-const kRunBreakResult = Symbol("runBreakResult");
-const kDocumentPosition = Symbol("documentPosition");
-const kInlineBlockRect = Symbol("inlineBlockRect");
-const kAbsolutePosition = Symbol("absolutePosition");
-const kSplitsAroundBlock = Symbol("splitsAroundBlock");
-const kBreakResultTextIndex = Symbol("breakResultTextIndex");
-const kRectTextIndices = Symbol("rectTextIndices");
-const kRangeTextNodes = Symbol("rangeTextNodes");
-const kCaretRect = Symbol("caretRect");
-const kSelectionRuns = Symbol("selectionRuns");
-const kOffsetInFragment = Symbol("offsetInFragment");
-const kHitTestContext = Symbol("hitTestContext");
-const kHitTestInFlow = Symbol("hitTestInFlow");
-const kPrincipalBox = Symbol("principalBox");
-const kStructuralGeneration = Symbol("structuralGeneration");
-const kStaleContainers = Symbol("staleContainers");
-const kFlowChildren = Symbol("flowChildren");
-const kIsSuppressedFlexWhitespace = Symbol("isSuppressedFlexWhitespace");
-const kRetireSteppedOver = Symbol("retireSteppedOver");
-const kInvalidationEpoch = Symbol("invalidationEpoch");
-const kMeasuresAsRun = Symbol("measuresAsRun");
-const kBoxEntryOf = Symbol("boxEntryOf");
-const kRunContainerOf = Symbol("runContainerOf");
-const kContainerBox = Symbol("containerBox");
-const kBoxParentOf = Symbol("boxParentOf");
-const kBoxOf = Symbol("boxOf");
-const kRetireHiddenContent = Symbol("retireHiddenContent");
-const kStaticPosition = Symbol("staticPosition");
-const kInlineCursorBefore = Symbol("inlineCursorBefore");
-const kHiddenByAncestor = Symbol("hiddenByAncestor");
-const kRetireRunContent = Symbol("retireRunContent");
-const kMeasureInlineRun = Symbol("measureInlineRun");
-const kSyncRunMembers = Symbol("syncRunMembers");
-const kRetireFlexNode = Symbol("retireFlexNode");
-const kRunContainerFrom = Symbol("runContainerFrom");
-const kRestageSubtree = Symbol("restageSubtree");
-const kInvalidateNode = Symbol("invalidateNode");
-const kRestageContainer = Symbol("restageContainer");
-const kRestageBox = Symbol("restageBox");
-const kRestageChildren = Symbol("restageChildren");
-const kStyleNode = Symbol("styleNode");
-const kClearBreakResultCache = Symbol("clearBreakResultCache");
-const kInvalidateNodeChildren = Symbol("invalidateNodeChildren");
-const kInvalidateBox = Symbol("invalidateBox");
-const kMarkRunMeasureDirty = Symbol("markRunMeasureDirty");
-const kEnclosingContentRoot = Symbol("enclosingContentRoot");
-const kInvalidateEnclosingMeasure = Symbol("invalidateEnclosingMeasure");
-const kInvalidateContainerBoxes = Symbol("invalidateContainerBoxes");
-const kRestageForRecord = Symbol("restageForRecord");
-const kRestyled = Symbol("restyled");
-const kContainingBlockFlexNode = Symbol("containingBlockFlexNode");
-const kAddElementNode = Symbol("addElementNode");
-const kAddTextNode = Symbol("addTextNode");
-const kBoxKindMatches = Symbol("boxKindMatches");
-const kSyncContentRoot = Symbol("syncContentRoot");
-const kContainsBlockLevelBox = Symbol("containsBlockLevelBox");
-const kTrackNode = Symbol("trackNode");
-const kRetireContentRoot = Symbol("retireContentRoot");
-const kRetireContainerBoxes = Symbol("retireContainerBoxes");
-const kBreakNodes = Symbol("breakNodes");
-const kCollectLeavesUnder = Symbol("collectLeavesUnder");
-const kShouldCollapseWhitespaceTextNode = Symbol(
-	"shouldCollapseWhitespaceTextNode",
-);
-const kIsRowFlexItem = Symbol("isRowFlexItem");
-const kFirstComposedRenderableChild = Symbol("firstComposedRenderableChild");
-const kCollectLeafNodes = Symbol("collectLeafNodes");
-const kProcessWhitespace = Symbol("processWhitespace");
-const kHasNowrapLeaf = Symbol("hasNowrapLeaf");
-const kFindBreakPoints = Symbol("findBreakPoints");
-const kBuildLines = Symbol("buildLines");
-const kRenderedLeaves = Symbol("renderedLeaves");
-const kRenderLeaf = Symbol("renderLeaf");
-const kMeasureText = Symbol("measureText");
-const kGetNodesInRange = Symbol("getNodesInRange");
-const kToVisualLine = Symbol("toVisualLine");
-
 /**
  * Whether a box takes part in positioned layout -- the predicate both the
  * containing-block chain and stacking-context collection are built on. Also
@@ -1323,6 +1229,100 @@ export {kInvalidateInlineRun};
 // consumer can render for itself. The layout tests reach it through the symbol.
 const kRectTexts = Symbol("rectTexts");
 export {kRectTexts};
+
+const kBoxes = Symbol("boxes");
+const kAnonymousBoxes = Symbol("anonymousBoxes");
+const kMeasureNodes = Symbol("measureNodes");
+const kTerminalReordersText = Symbol("terminalReordersText");
+const kDOMNodeByFlexNode = Symbol("domNodeByFlexNode");
+const kInvalidatedNodes = Symbol("invalidatedNodes");
+const kAddNode = Symbol("addNode");
+const kLayoutPass = Symbol("layoutPass");
+const kApplyRestyles = Symbol("applyRestyles");
+const kDirtyRunContainers = Symbol("dirtyRunContainers");
+const kPruneDisconnectedNodes = Symbol("pruneDisconnectedNodes");
+const kContainerFlexNode = Symbol("containerFlexNode");
+const kSyncContainerRuns = Symbol("syncContainerRuns");
+const kIsNodeLive = Symbol("isNodeLive");
+const kRetireAnonymousBox = Symbol("retireAnonymousBox");
+const kUntrackNode = Symbol("untrackNode");
+const kRunFlexNode = Symbol("runFlexNode");
+const kRunBreakResult = Symbol("runBreakResult");
+const kDocumentPosition = Symbol("documentPosition");
+const kInlineBlockRect = Symbol("inlineBlockRect");
+const kAbsolutePosition = Symbol("absolutePosition");
+const kSplitsAroundBlock = Symbol("splitsAroundBlock");
+const kBreakResultTextIndex = Symbol("breakResultTextIndex");
+const kRectTextIndices = Symbol("rectTextIndices");
+const kRangeTextNodes = Symbol("rangeTextNodes");
+const kCaretRect = Symbol("caretRect");
+const kSelectionRuns = Symbol("selectionRuns");
+const kOffsetInFragment = Symbol("offsetInFragment");
+const kHitTestContext = Symbol("hitTestContext");
+const kHitTestInFlow = Symbol("hitTestInFlow");
+const kPrincipalBox = Symbol("principalBox");
+const kStructuralGeneration = Symbol("structuralGeneration");
+const kStaleContainers = Symbol("staleContainers");
+const kFlowChildren = Symbol("flowChildren");
+const kIsSuppressedFlexWhitespace = Symbol("isSuppressedFlexWhitespace");
+const kRetireSteppedOver = Symbol("retireSteppedOver");
+const kInvalidationEpoch = Symbol("invalidationEpoch");
+const kMeasuresAsRun = Symbol("measuresAsRun");
+const kBoxEntryOf = Symbol("boxEntryOf");
+const kRunContainerOf = Symbol("runContainerOf");
+const kContainerBox = Symbol("containerBox");
+const kBoxParentOf = Symbol("boxParentOf");
+const kBoxOf = Symbol("boxOf");
+const kRetireHiddenContent = Symbol("retireHiddenContent");
+const kStaticPosition = Symbol("staticPosition");
+const kInlineCursorBefore = Symbol("inlineCursorBefore");
+const kHiddenByAncestor = Symbol("hiddenByAncestor");
+const kRetireRunContent = Symbol("retireRunContent");
+const kMeasureInlineRun = Symbol("measureInlineRun");
+const kSyncRunMembers = Symbol("syncRunMembers");
+const kRetireFlexNode = Symbol("retireFlexNode");
+const kRunContainerFrom = Symbol("runContainerFrom");
+const kRestageSubtree = Symbol("restageSubtree");
+const kInvalidateNode = Symbol("invalidateNode");
+const kRestageContainer = Symbol("restageContainer");
+const kRestageBox = Symbol("restageBox");
+const kRestageChildren = Symbol("restageChildren");
+const kStyleNode = Symbol("styleNode");
+const kClearBreakResultCache = Symbol("clearBreakResultCache");
+const kInvalidateNodeChildren = Symbol("invalidateNodeChildren");
+const kInvalidateBox = Symbol("invalidateBox");
+const kMarkRunMeasureDirty = Symbol("markRunMeasureDirty");
+const kEnclosingContentRoot = Symbol("enclosingContentRoot");
+const kInvalidateEnclosingMeasure = Symbol("invalidateEnclosingMeasure");
+const kInvalidateContainerBoxes = Symbol("invalidateContainerBoxes");
+const kRestageForRecord = Symbol("restageForRecord");
+const kRestyled = Symbol("restyled");
+const kContainingBlockFlexNode = Symbol("containingBlockFlexNode");
+const kAddElementNode = Symbol("addElementNode");
+const kAddTextNode = Symbol("addTextNode");
+const kBoxKindMatches = Symbol("boxKindMatches");
+const kSyncContentRoot = Symbol("syncContentRoot");
+const kContainsBlockLevelBox = Symbol("containsBlockLevelBox");
+const kTrackNode = Symbol("trackNode");
+const kRetireContentRoot = Symbol("retireContentRoot");
+const kRetireContainerBoxes = Symbol("retireContainerBoxes");
+const kBreakNodes = Symbol("breakNodes");
+const kCollectLeavesUnder = Symbol("collectLeavesUnder");
+const kShouldCollapseWhitespaceTextNode = Symbol(
+	"shouldCollapseWhitespaceTextNode",
+);
+const kIsRowFlexItem = Symbol("isRowFlexItem");
+const kFirstComposedRenderableChild = Symbol("firstComposedRenderableChild");
+const kCollectLeafNodes = Symbol("collectLeafNodes");
+const kProcessWhitespace = Symbol("processWhitespace");
+const kHasNowrapLeaf = Symbol("hasNowrapLeaf");
+const kFindBreakPoints = Symbol("findBreakPoints");
+const kBuildLines = Symbol("buildLines");
+const kRenderedLeaves = Symbol("renderedLeaves");
+const kRenderLeaf = Symbol("renderLeaf");
+const kMeasureText = Symbol("measureText");
+const kGetNodesInRange = Symbol("getNodesInRange");
+const kToVisualLine = Symbol("toVisualLine");
 
 export class LayoutEngine {
 	declare DOMRect: typeof DOMRect;
