+++ src/internal/termdom.ts
@@ -53,110 +53,6 @@ import {
 	tokenizeInput,
 } from "./events.js";
 
-const kWrite = Symbol("write");
-const kFullscreenStack = Symbol("fullscreenStack");
-const kIsInFullscreenMode = Symbol("isInFullscreenMode");
-const kEnterFullscreenMode = Symbol("enterFullscreenMode");
-const kFireFullscreenChangeEvent = Symbol("fireFullscreenChangeEvent");
-const kFireFullscreenErrorEvent = Symbol("fireFullscreenErrorEvent");
-const kExitFullscreenMode = Symbol("exitFullscreenMode");
-const kGetWindow = Symbol("getWindow");
-const kFrameDamage = Symbol("frameDamage");
-const kTransport = Symbol("transport");
-const kInteractive = Symbol("interactive");
-const kWidth = Symbol("width");
-const kHeight = Symbol("height");
-const kTopLayer = Symbol("topLayer");
-const kInstallPrototypes = Symbol("installPrototypes");
-const kScreen = Symbol("screen");
-const kStyleManager = Symbol("styleManager");
-const kProcessPendingMutationsAndRender = Symbol(
-	"processPendingMutationsAndRender",
-);
-const kFullscreenManager = Symbol("fullscreenManager");
-const kSession = Symbol("session");
-const kObserverManager = Symbol("observerManager");
-const kInstallWindowExtensions = Symbol("installWindowExtensions");
-const kInstallObservers = Symbol("installObservers");
-const kSetupMutationObserver = Symbol("setupMutationObserver");
-const kRender = Symbol("render");
-const kPainter = Symbol("painter");
-const kViewport = Symbol("viewport");
-const kBuildSession = Symbol("buildSession");
-const kOnFieldEditEvent = Symbol("onFieldEditEvent");
-const kOnDisclosureToggle = Symbol("onDisclosureToggle");
-const kQueueCaretReveal = Symbol("queueCaretReveal");
-const kNextRafId = Symbol("nextRafId");
-const kFlushDocument = Symbol("flushDocument");
-const kSealed = Symbol("sealed");
-const kScrollCamera = Symbol("scrollCamera");
-const kAllocateFrameHandle = Symbol("allocateFrameHandle");
-const kFrameCallbacks = Symbol("frameCallbacks");
-const kMediaQueryUpdaters = Symbol("mediaQueryUpdaters");
-const kAttached = Symbol("attached");
-const kAttachReady = Symbol("attachReady");
-const kRenderCount = Symbol("renderCount");
-const kSealToScrollback = Symbol("sealToScrollback");
-const kPrototypesInstalled = Symbol("prototypesInstalled");
-const kScreenSwitching = Symbol("screenSwitching");
-const kRenderInFlight = Symbol("renderInFlight");
-const kUpdateMouseReporting = Symbol("updateMouseReporting");
-const kFindElementAtDocumentPoint = Symbol("findElementAtDocumentPoint");
-const kAddFrameDamage = Symbol("addFrameDamage");
-const kHandlePendingMutations = Symbol("handlePendingMutations");
-const kInputGeneration = Symbol("inputGeneration");
-const kMouseCaptureYielded = Symbol("mouseCaptureYielded");
-const kReclaimMouseCapture = Symbol("reclaimMouseCapture");
-const kDispatchGlobalKeyboardEvent = Symbol("dispatchGlobalKeyboardEvent");
-const kHandleMouseReport = Symbol("handleMouseReport");
-const kDispatchPaste = Symbol("dispatchPaste");
-const kScheduleResize = Symbol("scheduleResize");
-const kRebindTransport = Symbol("rebindTransport");
-const kAttachBeginning = Symbol("attachBeginning");
-const kAttachBegun = Symbol("attachBegun");
-const kDisposed = Symbol("disposed");
-const kApplyTerminalSize = Symbol("applyTerminalSize");
-const kMouseReportingEnabled = Symbol("mouseReportingEnabled");
-const kScrollChainTimer = Symbol("scrollChainTimer");
-const kResizeInProgress = Symbol("resizeInProgress");
-const kIsRendering = Symbol("isRendering");
-const kRenderQueued = Symbol("renderQueued");
-const kRenderOnce = Symbol("renderOnce");
-const kDrainFrameCallbacks = Symbol("drainFrameCallbacks");
-const kRenderInteractive = Symbol("renderInteractive");
-const kPendingCaretReveal = Symbol("pendingCaretReveal");
-const kResizeEpoch = Symbol("resizeEpoch");
-const kResizeTimer = Symbol("resizeTimer");
-const kHandleResize = Symbol("handleResize");
-const kTopmostModalDialog = Symbol("topmostModalDialog");
-const kSCROLL_CHAIN_TIMEOUT_MS = Symbol("SCROLL_CHAIN_TIMEOUT_MS");
-const kFieldDragAnchor = Symbol("fieldDragAnchor");
-const kFieldOffsetAtPoint = Symbol("fieldOffsetAtPoint");
-const kSelectionDragAnchor = Symbol("selectionDragAnchor");
-const kMouseDownTarget = Symbol("mouseDownTarget");
-const kDocumentPointToTextPosition = Symbol("documentPointToTextPosition");
-const kPopoverPressTarget = Symbol("popoverPressTarget");
-const kLastClickTarget = Symbol("lastClickTarget");
-const kLastClickTime = Symbol("lastClickTime");
-const kDBLCLICK_INTERVAL_MS = Symbol("DBLCLICK_INTERVAL_MS");
-const kTopmostCloseRequestTarget = Symbol("topmostCloseRequestTarget");
-const kMoveFocus = Symbol("moveFocus");
-const kDispatchInsertText = Symbol("dispatchInsertText");
-const kAfterRender = Symbol("afterRender");
-const kScrollCaretIntoView = Symbol("scrollCaretIntoView");
-const kLastFrameScrollTop = Symbol("lastFrameScrollTop");
-const kLastFrameEpoch = Symbol("lastFrameEpoch");
-const kLastFrameInputGeneration = Symbol("lastFrameInputGeneration");
-const kLastFrameActiveElement = Symbol("lastFrameActiveElement");
-const kScrollFieldCaretIntoView = Symbol("scrollFieldCaretIntoView");
-const kDocumentPaintHeight = Symbol("documentPaintHeight");
-const kReserveRows = Symbol("reserveRows");
-const kLastFrameStructuralGeneration = Symbol("lastFrameStructuralGeneration");
-const kLastFrameSelectionLive = Symbol("lastFrameSelectionLive");
-const kStaticSibling = Symbol("staticSibling");
-const kRenderStaticHTML = Symbol("renderStaticHTML");
-const kStaticRenderer = Symbol("staticRenderer");
-
 // How long to wait for a resize drag to settle before redrawing. Long enough to
 // coalesce the burst of SIGWINCHes a drag fires, short enough to feel immediate.
 const RESIZE_DEBOUNCE_MS = 40;
@@ -242,8 +138,16 @@ const kObserver = Symbol("observer");
 // The static-render entry renderANSI() reaches through; off the public API
 // like the test handles above.
 const kRenderStatic = Symbol("renderStatic");
-const kPrintStatic = Symbol("printStatic");
 export {kLayoutEngine, kObserver, kRenderStatic};
+
+const kWrite = Symbol("write");
+const kFullscreenStack = Symbol("fullscreenStack");
+const kIsInFullscreenMode = Symbol("isInFullscreenMode");
+const kEnterFullscreenMode = Symbol("enterFullscreenMode");
+const kFireFullscreenChangeEvent = Symbol("fireFullscreenChangeEvent");
+const kFireFullscreenErrorEvent = Symbol("fireFullscreenErrorEvent");
+const kExitFullscreenMode = Symbol("exitFullscreenMode");
+const kGetWindow = Symbol("getWindow");
 
 /**
  * The Fullscreen API over the terminal's alternate screen. Lives here
@@ -610,6 +514,103 @@ function createEngineWindow(document: DOM.Document): EngineWindow {
 	DOM.setAmbientDocument(document);
 	return window as unknown as EngineWindow;
 }
+
+const kFrameDamage = Symbol("frameDamage");
+const kTransport = Symbol("transport");
+const kInteractive = Symbol("interactive");
+const kWidth = Symbol("width");
+const kHeight = Symbol("height");
+const kTopLayer = Symbol("topLayer");
+const kInstallPrototypes = Symbol("installPrototypes");
+const kScreen = Symbol("screen");
+const kStyleManager = Symbol("styleManager");
+const kProcessPendingMutationsAndRender = Symbol(
+	"processPendingMutationsAndRender",
+);
+const kFullscreenManager = Symbol("fullscreenManager");
+const kSession = Symbol("session");
+const kObserverManager = Symbol("observerManager");
+const kInstallWindowExtensions = Symbol("installWindowExtensions");
+const kInstallObservers = Symbol("installObservers");
+const kSetupMutationObserver = Symbol("setupMutationObserver");
+const kRender = Symbol("render");
+const kPainter = Symbol("painter");
+const kViewport = Symbol("viewport");
+const kBuildSession = Symbol("buildSession");
+const kOnFieldEditEvent = Symbol("onFieldEditEvent");
+const kOnDisclosureToggle = Symbol("onDisclosureToggle");
+const kQueueCaretReveal = Symbol("queueCaretReveal");
+const kNextRafId = Symbol("nextRafId");
+const kFlushDocument = Symbol("flushDocument");
+const kSealed = Symbol("sealed");
+const kScrollCamera = Symbol("scrollCamera");
+const kAllocateFrameHandle = Symbol("allocateFrameHandle");
+const kFrameCallbacks = Symbol("frameCallbacks");
+const kMediaQueryUpdaters = Symbol("mediaQueryUpdaters");
+const kAttached = Symbol("attached");
+const kAttachReady = Symbol("attachReady");
+const kRenderCount = Symbol("renderCount");
+const kSealToScrollback = Symbol("sealToScrollback");
+const kPrototypesInstalled = Symbol("prototypesInstalled");
+const kScreenSwitching = Symbol("screenSwitching");
+const kRenderInFlight = Symbol("renderInFlight");
+const kUpdateMouseReporting = Symbol("updateMouseReporting");
+const kFindElementAtDocumentPoint = Symbol("findElementAtDocumentPoint");
+const kAddFrameDamage = Symbol("addFrameDamage");
+const kHandlePendingMutations = Symbol("handlePendingMutations");
+const kInputGeneration = Symbol("inputGeneration");
+const kMouseCaptureYielded = Symbol("mouseCaptureYielded");
+const kReclaimMouseCapture = Symbol("reclaimMouseCapture");
+const kDispatchGlobalKeyboardEvent = Symbol("dispatchGlobalKeyboardEvent");
+const kHandleMouseReport = Symbol("handleMouseReport");
+const kDispatchPaste = Symbol("dispatchPaste");
+const kScheduleResize = Symbol("scheduleResize");
+const kRebindTransport = Symbol("rebindTransport");
+const kAttachBeginning = Symbol("attachBeginning");
+const kAttachBegun = Symbol("attachBegun");
+const kDisposed = Symbol("disposed");
+const kApplyTerminalSize = Symbol("applyTerminalSize");
+const kMouseReportingEnabled = Symbol("mouseReportingEnabled");
+const kScrollChainTimer = Symbol("scrollChainTimer");
+const kResizeInProgress = Symbol("resizeInProgress");
+const kIsRendering = Symbol("isRendering");
+const kRenderQueued = Symbol("renderQueued");
+const kRenderOnce = Symbol("renderOnce");
+const kDrainFrameCallbacks = Symbol("drainFrameCallbacks");
+const kRenderInteractive = Symbol("renderInteractive");
+const kPendingCaretReveal = Symbol("pendingCaretReveal");
+const kResizeEpoch = Symbol("resizeEpoch");
+const kResizeTimer = Symbol("resizeTimer");
+const kHandleResize = Symbol("handleResize");
+const kTopmostModalDialog = Symbol("topmostModalDialog");
+const kSCROLL_CHAIN_TIMEOUT_MS = Symbol("SCROLL_CHAIN_TIMEOUT_MS");
+const kFieldDragAnchor = Symbol("fieldDragAnchor");
+const kFieldOffsetAtPoint = Symbol("fieldOffsetAtPoint");
+const kSelectionDragAnchor = Symbol("selectionDragAnchor");
+const kMouseDownTarget = Symbol("mouseDownTarget");
+const kDocumentPointToTextPosition = Symbol("documentPointToTextPosition");
+const kPopoverPressTarget = Symbol("popoverPressTarget");
+const kLastClickTarget = Symbol("lastClickTarget");
+const kLastClickTime = Symbol("lastClickTime");
+const kDBLCLICK_INTERVAL_MS = Symbol("DBLCLICK_INTERVAL_MS");
+const kTopmostCloseRequestTarget = Symbol("topmostCloseRequestTarget");
+const kMoveFocus = Symbol("moveFocus");
+const kDispatchInsertText = Symbol("dispatchInsertText");
+const kAfterRender = Symbol("afterRender");
+const kScrollCaretIntoView = Symbol("scrollCaretIntoView");
+const kLastFrameScrollTop = Symbol("lastFrameScrollTop");
+const kLastFrameEpoch = Symbol("lastFrameEpoch");
+const kLastFrameInputGeneration = Symbol("lastFrameInputGeneration");
+const kLastFrameActiveElement = Symbol("lastFrameActiveElement");
+const kScrollFieldCaretIntoView = Symbol("scrollFieldCaretIntoView");
+const kDocumentPaintHeight = Symbol("documentPaintHeight");
+const kReserveRows = Symbol("reserveRows");
+const kLastFrameStructuralGeneration = Symbol("lastFrameStructuralGeneration");
+const kLastFrameSelectionLive = Symbol("lastFrameSelectionLive");
+const kStaticSibling = Symbol("staticSibling");
+const kRenderStaticHTML = Symbol("renderStaticHTML");
+const kStaticRenderer = Symbol("staticRenderer");
+const kPrintStatic = Symbol("printStatic");
 
 /**
  * Everything the window installers below need from the TermDOM that installed
